add styles.css and add some float experimentation
A lot of practice with css and a long time without a commit. I added
styles to all the pages and also started experimenting with floats at
the bottom of the lasagna page. This will be eventaully be deleted but
I'm committing it, instead of deleting it, so I can eventually go back
and review the float material I learned; I will eventually need to.